# models/nick_name.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/strfmt"
)

// NickName nick name
//
// swagger:model nickName
type NickName string

// Validate validates this nick name
func (m NickName) Validate(formats strfmt.Registry) error {
	return nil
}