# models/user_data_filter.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// UserDataFilter The porpouse of this object is to pass filtering data to the query
//
// swagger:model userDataFilter
type UserDataFilter struct {

	// include elements with these values
	Exclude *UserDataNoPassword `json:"exclude,omitempty"`

	// exclude elements with these values
	Include *UserDataNoPassword `json:"include,omitempty"`
}

// Validate validates this user data filter
func (m *UserDataFilter) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateExclude(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateInclude(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *UserDataFilter) validateExclude(formats strfmt.Registry) error {

	if swag.IsZero(m.Exclude) { // not required
		return nil
	}

	if m.Exclude != nil {
		if err := m.Exclude.Validate(formats); err != nil {
			if ve, ok := err.(*errors.Validation); ok {
				return ve.ValidateName("exclude")
			}
			return err
		}
	}

	return nil
}

func (m *UserDataFilter) validateInclude(formats strfmt.Registry) error {

	if swag.IsZero(m.Include) { // not required
		return nil
	}

	if m.Include != nil {
		if err := m.Include.Validate(formats); err != nil {
			if ve, ok := err.(*errors.Validation); ok {
				return ve.ValidateName("include")
			}
			return err
		}
	}

	return nil
}

// MarshalBinary interface implementation
func (m *UserDataFilter) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *UserDataFilter) UnmarshalBinary(b []byte) error {
	var res UserDataFilter
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}