# models/user_data_no_password.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// UserDataNoPassword This object is a copy of userData without the password
//
// swagger:model userDataNoPassword
type UserDataNoPassword struct {

	// country
	Country string `json:"country,omitempty"`

	// email
	Email string `json:"email,omitempty"`

	// firstName
	FirstName string `json:"firstName,omitempty"`

	// nickName
	NickName string `json:"nickName,omitempty"`

	// secondName
	SecondName string `json:"secondName,omitempty"`
}

// Validate validates this user data no password
func (m *UserDataNoPassword) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *UserDataNoPassword) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *UserDataNoPassword) UnmarshalBinary(b []byte) error {
	var res UserDataNoPassword
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}